Fix a launcher unresponsive detection race between virt-handler and the kubelet
Once all processes in the virt-launcher pod are gone, the kubelet will
start unmounting all volumes. One of these is the `sockets` volume.
virt-handler on the other hand uses this volume to mark launchers as
unresponsive. If virt-handler tries to mark a launcher as unresponsive
after the unmount of the sockets volume, it has to consider that this
directory may not exist anymore and implicitly assume that the launcher
is not responsive anymore at this point.
This is safe. There is no risk that virt-launcher is in a startup phase
at this moment, since virt-handler only ever acts on launcher which were
ready in the past.

Without this, virt-handler may run repeatedly into this error:

```
{"component":"virt-handler","level":"error","msg":"Unable to mark vmi as unresponsive socket //pods/e6cc43c8-dbde-4af3-96c8-e6ab88fabb07/volumes/kubernetes.io~empty-dir/sockets/launcher-sock","pos":"cache.go:485","reason":"open /pods/e6cc43c8-dbde-4af3-96c8-e6ab88fabb07/volumes/kubernetes.io~empty-dir/sockets/launcher-unresponsive: no such file or directory","timestamp":"2022-07-14T06:11:59.919964Z"}
```

and the VMI cleanup can be significantly delayed.

+++ pkg/virt-handler/cmd-client/client.go
@@ -196,6 +196,12 @@ func SocketMonitoringEnabled(socket string) bool {
 }
 
 func IsSocketUnresponsive(socket string) bool {
+	if exists, _ := diskutils.FileExists(filepath.Dir(socket)); !exists {
+		// If the sockets directory does not exist the kubelet already started unmounting
+		// since the sockets directory is a mountpoint.
+		return true
+	}
+
 	file := filepath.Join(filepath.Dir(socket), StandardLauncherUnresponsiveFileName)
 	exists, _ := diskutils.FileExists(file)
 	// if the unresponsive socket monitor marked this socket
@@ -214,6 +220,11 @@ func IsSocketUnresponsive(socket string) bool {
 }
 
 func MarkSocketUnresponsive(socket string) error {
+	if exists, _ := diskutils.FileExists(filepath.Dir(socket)); !exists {
+		// If the sockets directory does not exist the kubelet already started unmounting
+		// since the sockets directory is a mountpoint. Don't try to create a file at an inexistent location.
+		return nil
+	}
 	file := filepath.Join(filepath.Dir(socket), StandardLauncherUnresponsiveFileName)
 	f, err := os.Create(file)
 	if err != nil {
